moved epoch tracker tests
